refactor(test): consolidate per-endpoint extra tests into replication.test.ts
Adds three stub-based describe block inside the existing replication test
file rather than a separate file:
- slave extra merged over top-level extra
- credentials.extra takes priority when the same key is set in both
- backward-compat when credentials.extra is absent

Deletes the standalone replication-extra.test.ts introduced earlier.
